==> 04_Carcinoid_immune analyses.Rmd <==

```{r setup}

### load libraries
library(Seurat)
library(dplyr)
library(reticulate)
library(sctransform)
library(cowplot)
library(ggplot2)
library(viridis)
library(tidyr)
library(magrittr)
library(reshape2)
library(readxl)
library(stringr)
library(cowplot)
library(scales)
library(readr)
library(progeny)
library(gplots)
library(tibble)
library(grid)
library(rlang)
library(SeuratDisk)
library(ggsankey)

theme_set(theme_cowplot())

use_colors <- c(
  Normal = "deepskyblue2",
  Tumor = "brown2",
  G1 = "#46ACC8",
  G2M = "#E58601",
  S = "#B40F20",
  Epithelial = "seagreen",
  Immune = "darkgoldenrod2",
  Stromal = "steelblue",
  Patient1 = "#E2D200",
  Patient2 = "#FD6467",
  Patient3 = "#B40F20",
  Patient4 = "royalblue4",
  Patient5 = "dodgerblue3",
  Patient6 = "deepskyblue",
  Patient7 = "cadetblue1",
  Patient8 = "darkolivegreen2",
  Patient9 = "chartreuse3",
  Patient10 = "darkgreen",
  Macrophage = "#08519C",
  Macrophage_proliferating = "#6bAEd6",
  Myeloid = "#D40C00",
  Monocyte = "#FFCD00",
  cDC = "#EB96EB",
  pDC = "#7A0177",
  T_conv = "#bcbddc",
  T_CD8 = "#807dba",
  `T/NK_proliferating` = "#78c679",
  NK = "#006837",
  B = "#969696",
  Plasma = "#636363",
  Mast = "#252525")

```

```{r heatmap wrapper function}

###code from https://github.com/satijalab/seurat/issues/2201

DoMultiBarHeatmap <- function (object, 
                               features = NULL, 
                               cells = NULL, 
                               group.by = "ident", 
                               additional.group.by = NULL, 
                               additional.group.sort.by = NULL, 
                               cols.use = NULL,
                               group.bar = TRUE, 
                               disp.min = -2.5, 
                               disp.max = NULL, 
                               slot = "scale.data", 
                               assay = NULL, 
                               label = TRUE, 
                               size = 5.5, 
                               hjust = 0, 
                               angle = 45, 
                               raster = TRUE, 
                               draw.lines = TRUE, 
                               lines.width = NULL, 
                               group.bar.height = 0.02, 
                               combine = TRUE) 
{
  cells <- cells %||% colnames(x = object)
  if (is.numeric(x = cells)) {
    cells <- colnames(x = object)[cells]
  }
  assay <- assay %||% DefaultAssay(object = object)
  DefaultAssay(object = object) <- assay
  features <- features %||% VariableFeatures(object = object)
  ## Why reverse???
  features <- rev(x = unique(x = features))
  disp.max <- disp.max %||% ifelse(test = slot == "scale.data", 
                                   yes = 2.5, no = 6)
  possible.features <- rownames(x = GetAssayData(object = object, 
                                                 slot = slot))
  if (any(!features %in% possible.features)) {
    bad.features <- features[!features %in% possible.features]
    features <- features[features %in% possible.features]
    if (length(x = features) == 0) {
      stop("No requested features found in the ", slot, 
           " slot for the ", assay, " assay.")
    }
    warning("The following features were omitted as they were not found in the ", 
            slot, " slot for the ", assay, " assay: ", paste(bad.features, 
                                                             collapse = ", "))
  }
  
  if (!is.null(additional.group.sort.by)) {
    if (any(!additional.group.sort.by %in% additional.group.by)) {
      bad.sorts <- additional.group.sort.by[!additional.group.sort.by %in% additional.group.by]
      additional.group.sort.by <- additional.group.sort.by[additional.group.sort.by %in% additional.group.by]
      if (length(x = bad.sorts) > 0) {
        warning("The following additional sorts were omitted as they were not a subset of additional.group.by : ", 
                paste(bad.sorts, collapse = ", "))
      }
    }
  }
  
  data <- as.data.frame(x = as.matrix(x = t(x = GetAssayData(object = object, 
                                                             slot = slot)[features, cells, drop = FALSE])))
  
  object <- suppressMessages(expr = StashIdent(object = object, 
                                               save.name = "ident"))
  group.by <- group.by %||% "ident"
  groups.use <- object[[c(group.by, additional.group.by[!additional.group.by %in% group.by])]][cells, , drop = FALSE]
  plots <- list()
  for (i in group.by) {
    data.group <- data
    if (!is_null(additional.group.by)) {
      additional.group.use <- additional.group.by[additional.group.by!=i]  
      if (!is_null(additional.group.sort.by)){
        additional.sort.use = additional.group.sort.by[additional.group.sort.by != i]  
      } else {
        additional.sort.use = NULL
      }
    } else {
      additional.group.use = NULL
      additional.sort.use = NULL
    }
    
    group.use <- groups.use[, c(i, additional.group.use), drop = FALSE]
    
    for(colname in colnames(group.use)){
      if (!is.factor(x = group.use[[colname]])) {
        group.use[[colname]] <- factor(x = group.use[[colname]])
      }  
    }
    
    if (draw.lines) {
      lines.width <- lines.width %||% ceiling(x = nrow(x = data.group) * 
                                                0.0025)
      placeholder.cells <- sapply(X = 1:(length(x = levels(x = group.use[[i]])) * 
                                           lines.width), FUN = function(x) {
                                             return(Seurat:::RandomName(length = 20))
                                           })
      placeholder.groups <- data.frame(rep(x = levels(x = group.use[[i]]), times = lines.width))
      group.levels <- list()
      group.levels[[i]] = levels(x = group.use[[i]])
      for (j in additional.group.use) {
        group.levels[[j]] <- levels(x = group.use[[j]])
        placeholder.groups[[j]] = NA
      }
      
      colnames(placeholder.groups) <- colnames(group.use)
      rownames(placeholder.groups) <- placeholder.cells
      
      group.use <- sapply(group.use, as.vector)
      rownames(x = group.use) <- cells
      
      group.use <- rbind(group.use, placeholder.groups)
      
      for (j in names(group.levels)) {
        group.use[[j]] <- factor(x = group.use[[j]], levels = group.levels[[j]])
      }
      
      na.data.group <- matrix(data = NA, nrow = length(x = placeholder.cells), 
                              ncol = ncol(x = data.group), dimnames = list(placeholder.cells, 
                                                                           colnames(x = data.group)))
      data.group <- rbind(data.group, na.data.group)
    }
    
    order_expr <- paste0('order(', paste(c(i, additional.sort.use), collapse=','), ')')
    group.use = with(group.use, group.use[eval(parse(text=order_expr)), , drop=F])
    
    plot <- Seurat:::SingleRasterMap(data = data.group, raster = raster, 
                                     disp.min = disp.min, disp.max = disp.max, feature.order = features, 
                                     cell.order = rownames(x = group.use), group.by = group.use[[i]])
    
    if (group.bar) {
      pbuild <- ggplot_build(plot = plot)
      group.use2 <- group.use
      cols <- list()
      na.group <- Seurat:::RandomName(length = 20)
      for (colname in rev(x = colnames(group.use2))) {
        if (colname == i) {
          colid = paste0('Identity (', colname, ')')
        } else {
          colid = colname
        }
        
        # Default
        cols[[colname]] <- c(scales::hue_pal()(length(x = levels(x = group.use[[colname]]))))  
        
        #Overwrite if better value is provided
        if (!is_null(cols.use[[colname]])) {
          req_length = length(x = levels(group.use))
          if (length(cols.use[[colname]]) < req_length){
            warning("Cannot use provided colors for ", colname, " since there aren't enough colors.")
          } else {
            if (!is_null(names(cols.use[[colname]]))) {
              if (all(levels(group.use[[colname]]) %in% names(cols.use[[colname]]))) {
                cols[[colname]] <- as.vector(cols.use[[colname]][levels(group.use[[colname]])])
              } else {
                warning("Cannot use provided colors for ", colname, " since all levels (", paste(levels(group.use[[colname]]), collapse=","), ") are not represented.")
              }
            } else {
              cols[[colname]] <- as.vector(cols.use[[colname]])[c(1:length(x = levels(x = group.use[[colname]])))]
            }
          }
        }
        
        # Add white if there's lines
        if (draw.lines) {
          levels(x = group.use2[[colname]]) <- c(levels(x = group.use2[[colname]]), na.group)  
          group.use2[placeholder.cells, colname] <- na.group
          cols[[colname]] <- c(cols[[colname]], "#FFFFFF")
        }
        names(x = cols[[colname]]) <- levels(x = group.use2[[colname]])
        
        y.range <- diff(x = pbuild$layout$panel_params[[1]]$y.range)
        y.pos <- max(pbuild$layout$panel_params[[1]]$y.range) + y.range * 0.015
        y.max <- y.pos + group.bar.height * y.range
        pbuild$layout$panel_params[[1]]$y.range <- c(pbuild$layout$panel_params[[1]]$y.range[1], y.max)
        
        plot <- suppressMessages(plot + 
                                   annotation_raster(raster = t(x = cols[[colname]][group.use2[[colname]]]),  xmin = -Inf, xmax = Inf, ymin = y.pos, ymax = y.max) + 
                                   annotation_custom(grob = grid::textGrob(label = colid, hjust = 0, gp = gpar(cex = 0.75)), ymin = mean(c(y.pos, y.max)), ymax = mean(c(y.pos, y.max)), xmin = Inf, xmax = Inf) +
                                   coord_cartesian(ylim = c(0, y.max), clip = "off")) 
        
        if ((colname == i) && label) {
          x.max <- max(pbuild$layout$panel_params[[1]]$x.range)
          x.divs <- pbuild$layout$panel_params[[1]]$x.major %||% pbuild$layout$panel_params[[1]]$x$break_positions()
          group.use$x <- x.divs
          label.x.pos <- tapply(X = group.use$x, INDEX = group.use[[colname]],
                                FUN = median) * x.max
          label.x.pos <- data.frame(group = names(x = label.x.pos), 
                                    label.x.pos)
          plot <- plot + geom_text(stat = "identity", 
                                   data = label.x.pos, aes_string(label = "group", 
                                                                  x = "label.x.pos"), y = y.max + y.max * 
                                     0.03 * 0.5, angle = angle, hjust = hjust, 
                                   size = size)
          plot <- suppressMessages(plot + coord_cartesian(ylim = c(0, 
                                                                   y.max + y.max * 0.002 * max(nchar(x = levels(x = group.use[[colname]]))) * 
                                                                     size), clip = "off"))
        }
      }
    }
    plot <- plot + theme(line = element_blank())
    plots[[i]] <- plot
  }
  if (combine) {
    plots <- CombinePlots(plots = plots)
  }
  return(plots)
}
```

```{r load data}

imm_anno <- readRDS("seurat_objects/imm_anno.RDS")

imm_anno@meta.data$cell_type_imm <- ordered(imm_anno@meta.data$cell_type_imm, levels = c("Monocyte",
                                                                                         "Myeloid",
                                                                                         "Macrophage",
                                                                                         "Macrophage_proliferating",
                                                                                         "cDC",
                                                                                         "pDC",
                                                                                         "Mast",
                                                                                         "T_conv",
                                                                                         "T_CD8",
                                                                                         "NK",
                                                                                         "T/NK_proliferating",
                                                                                         "B",
                                                                                         "Plasma"))

```

```{r UMAPs and marker genes}

DimPlot(imm_anno, group.by = "tissue_type", cols = use_colors)
ggsave2("DimPlot_imm_Normal_Tumor.pdf", path = "output/fig3", width = 15, height = 15, units = "cm")
ggsave2("DimPlot_imm_Normal_Tumor.png", path = "output/fig3", width = 35, height = 15, units = "cm")

DimPlot(imm_anno, group.by = "patient_id", cols = use_colors)
ggsave2("DimPlot_imm_patients.pdf", path = "output/fig3", width = 30, height = 15, units = "cm")
ggsave2("DimPlot_imm_patients.png", path = "output/fig3", width = 30, height = 15, units = "cm")

DimPlot(imm_anno, group.by = "cell_type_imm", split.by = "tissue_type", cols = use_colors, pt.size = 0.5)
ggsave2("DimPlot_imm_celltype.pdf", path = "output/fig3", width = 30, height = 15, units = "cm")
ggsave2("DimPlot_imm_celltype.png", path = "output/fig3", width = 30, height = 15, units = "cm")

DotPlot(imm_anno, features = c("S100A12", "FCN1", "LGMN", "CSF1R", "CD14", "CD68", "LYZ", "FABP4", "MARCO", "CD1C", "FCER1A", "LILRA4", "IL3RA", "KIT", "GATA2", "CD3E", "FOXP3", "CD8A", "NKG7", "KLRD1", "MS4A1", "CD79A", "JCHAIN", "IGKC", "MKI67"), group.by = "cell_type_imm") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  coord_flip() + 
  scale_color_viridis()
ggsave2("DotPlot_markergenes_imm.pdf", path = "output/fig3", width = 20, height = 20, units = "cm")


DotPlot(imm_anno, features = c("S100A12", "FCN1", "LGMN", "CSF1R", "CD14", "FCGR3A", "CD68", "LYZ", "FABP4", "MARCO", "CD1C", "FCER1A", "LILRA4", "IL3RA", "KIT", "GATA2", "CD3E", "FOXP3", "CD8A", "NKG7", "KLRD1", "MS4A1", "CD79A", "JCHAIN", "IGKC", "MKI67"), group.by = "cell_type_imm_refined") + 
  theme(axis.text.x = element_text(angle = 90, hjust = 1)) +
  coord_flip() + 
  scale_color_viridis()
```

```{r cell counts}

###subset data

imm_lympho <- subset(imm_anno, subset = cell_type_imm %in% c("T_conv",
                                                             "T_CD8",
                                                             "NK",
                                                             "T/NK_proliferating",
                                                             "B",
                                                             "Plasma"))

imm_lympho@meta.data$cell_type_imm_refined <- ordered(imm_lympho@meta.data$cell_type_imm_refined, levels = c("T_conv_1",
                                                                                                             "T_conv_2",
                                                                                                             "T_conv_3",
                                                                                                             "T_conv_4",
                                                                                                             "T_CD8_1",
                                                                                                             "T_CD8_2",
                                                                                                             "T_CD8_3",
                                                                                                             "T_CD8_4",
                                                                                                             "NK_1",
                                                                                                             "NK_2",
                                                                                                             "NK_3",
                                                                                                             "NK_4",
                                                                                                             "T/NK_proliferating",
                                                                                                             "B",
                                                                                                             "Plasma"))

imm_lympho <- ScaleData(imm_lympho)

imm_myelo <- subset(imm_anno, subset = cell_type_imm %in% c("Macrophage",
                                                            "Macrophage_proliferating",
                                                            "Myeloid",
                                                            "Monocyte",
                                                            "cDC",
                                                            "pDC",
                                                            "Mast"))

imm_myelo@meta.data$cell_type_imm_refined <- ordered(imm_myelo@meta.data$cell_type_imm_refined, levels = c("Monocyte_1",
                                                                                                           "Monocyte_2",
                                                                                                           "Monocyte_3",
                                                                                                           "Monocyte_4",
                                                                                                           "Myeloid_1",
                                                                                                           "Myeloid_2",
                                                                                                           "Myeloid_3",
                                                                                                           "Macrophage_1",
                                                                                                           "Macrophage_2",
                                                                                                           "Macrophage_3",
                                                                                                           "Macrophage_4",
                                                                                                           "Macrophage_5",
                                                                                                           "Macrophage_6",
                                                                                                           "Macrophage_7",
                                                                                                           "Macrophage_8",
                                                                                                           "Macrophage_9",
                                                                                                           "Macrophage_10",
                                                                                                           "Macrophage_11",
                                                                                                           "Macrophage_12",
                                                                                                           "Macrophage_13",
                                                                                                           "Macrophage_proliferating",
                                                                                                           "cDC_1",
                                                                                                           "cDC_2",
                                                                                                           "pDC",
                                                                                                           "Mast"))

imm_myelo <- ScaleData(imm_myelo)



###absolute cell counts

lympho_counts <- FetchData(imm_lympho, vars = c("tissue_type", "cell_type_imm", "cell_type_imm_refined", "sample_id", "patient_id")) %>%  
  mutate(tissue_type = factor(tissue_type, levels = c("Tumor", "Normal")))

lympho_counts_tbl <- lympho_counts %>%
  dplyr::count(cell_type_imm, patient_id, tissue_type)
write_csv(lympho_counts_tbl, path = "output/tables/lympho_counts.csv")

lympho_counts_tbl_refined <- lympho_counts %>%
  dplyr::count(cell_type_imm_refined, patient_id, tissue_type)
write_csv(lympho_counts_tbl_refined, path = "output/tables/lympho_counts_refined.csv")

myelo_counts <- FetchData(imm_myelo, vars = c("tissue_type", "cell_type_imm", "cell_type_imm_refined", "sample_id", "patient_id")) %>%  
  mutate(tissue_type = factor(tissue_type, levels = c("Tumor", "Normal"))) 

myelo_counts_tbl <- myelo_counts %>%
  dplyr::count(cell_type_imm, patient_id, tissue_type)
write_csv(myelo_counts_tbl, path = "output/tables/myelo_counts.csv")

myelo_counts_tbl_refined <- myelo_counts %>%
  dplyr::count(cell_type_imm_refined, patient_id, tissue_type)
write_csv(myelo_counts_tbl_refined, path = "output/tables/myelo_counts_refined.csv")


###plots cell counts

ggplot(data = lympho_counts, aes(x = tissue_type, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_lympho.pdf", path = "output/fig3", width = 20, height = 5, units = "cm")

ggplot(data = myelo_counts, aes(x = tissue_type, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_myelo.pdf", path = "output/fig3", width = 20, height = 5, units = "cm")

lympho_counts %>%
  filter(tissue_type == "Tumor") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_lympho_bypatient.pdf", path = "output/fig3", width = 30, height = 30, units = "cm")

myelo_counts %>%
  filter(tissue_type == "Tumor") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_myelo_bypatient.pdf", path = "output/fig3", width = 30, height = 30, units = "cm")

lympho_counts %>%
  filter(tissue_type == "Normal") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_lympho_bypatient_normal.pdf", path = "output/fig3", width = 30, height = 30, units = "cm")

myelo_counts %>%
  filter(tissue_type == "Normal") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_myelo_bypatient_normal.pdf", path = "output/fig3", width = 30, height = 30, units = "cm")



###relative cell counts

lympho_counts_rel <- lympho_counts %>%
  dplyr::count(cell_type_imm_refined, tissue_type) %>%
  pivot_wider(names_from =  tissue_type, values_from = n) %>%
  mutate(Tumor = ifelse(is.na(Tumor), 0,  Tumor)) %>%
  mutate(Normal = ifelse(is.na(Normal), 0,  Normal)) %>%
  pivot_longer(cols = c(Normal, Tumor), names_to = "tissue_type", values_to = "n") %>%
  group_by(tissue_type) %>% 
  mutate(n_rel = n/sum(n))

myelo_counts_rel <- myelo_counts %>%
  dplyr::count(cell_type_imm_refined, tissue_type) %>%
  pivot_wider(names_from =  tissue_type, values_from = n) %>%
  mutate(Tumor = ifelse(is.na(Tumor), 0,  Tumor)) %>%
  mutate(Normal = ifelse(is.na(Normal), 0,  Normal)) %>%
  pivot_longer(cols = c(Normal, Tumor), names_to = "tissue_type", values_to = "n") %>%
  group_by(tissue_type) %>% 
  mutate(n_rel = n/sum(n)) 

cDC_counts_rel <- myelo_counts_rel %>%
  filter(cell_type_imm_refined %in% c("cDC_1", "cDC_2"))

Monocytes_counts_rel <- myelo_counts_rel %>%
  filter(cell_type_imm_refined %in% c("Monocyte_1", "Monocyte_2", "Monocyte_3", "Monocyte_4"))



###plots relative cell counts

ggplot(lympho_counts_rel, aes(fill = tissue_type, y = n_rel, x = cell_type_imm_refined)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.7) +
  scale_fill_manual(values = use_colors)
ggsave2("BarPlot_lympho_refined_relative.pdf", path = "output/fig3", width = 30, height = 10, units = "cm")

ggplot(myelo_counts_rel, aes(fill = tissue_type, y = n_rel, x = cell_type_imm_refined)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.7) +
  scale_fill_manual(values = use_colors)
ggsave2("BarPlot_myelo_refined_relative.pdf", path = "output/fig3", width = 30, height = 10, units = "cm")

ggplot(cDC_counts_rel, aes(fill = tissue_type, y = n_rel, x = cell_type_imm_refined)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.7) +
  scale_fill_manual(values = use_colors)
ggsave2("BarPlot_cDC_refined_relative.pdf", path = "output/fig3", width = 10, height = 10, units = "cm")

ggplot(Monocytes_counts_rel, aes(fill = tissue_type, y = n_rel, x = cell_type_imm_refined)) +
  geom_bar(stat = "identity", position = "dodge", width = 0.7) +
  scale_fill_manual(values = use_colors)
ggsave2("BarPlot_Monocytes_refined_relative.pdf", path = "output/fig3", width = 20, height = 10, units = "cm")

```

```{r differential gene expression myeloid cells}

###all myeloid cells

Idents(imm_myelo) <- imm_myelo@meta.data$cell_type_imm

myelo_markers <- FindAllMarkers(imm_myelo, only.pos = T, min.pct = 0.25, min.diff.pct = 0.25)

top_myelo_markers <- myelo_markers %>% group_by(cluster) %>% top_n(10, wt = avg_log2FC)

DoMultiBarHeatmap(imm_myelo, features = top_myelo_markers$gene, group.by = "cell_type_imm", additional.group.by = "tissue_type",additional.group.sort.by = "tissue_type", cols.use = list(tissue_type = use_colors), draw.lines = F) +
  scale_fill_viridis()
ggsave2("HeatMap_Myelo_all.pdf", path = "output/fig3", width = 30, height = 40, units = "cm")
ggsave2("HeatMap_Myelo_all.png", path = "output/fig3", width = 30, height = 40, units = "cm")



###only myeloid cells in tumor samples

imm_myelo_subset1 <- subset(imm_myelo, tissue_type == "Tumor")

Idents(imm_myelo_subset1) <- imm_myelo_subset1@meta.data$cell_type_imm

myelo_markers <- FindAllMarkers(imm_myelo_subset1, only.pos = T, min.pct = 0.25, min.diff.pct = 0.25)

top_myelo_markers <- myelo_markers %>% group_by(cluster) %>% top_n(10, wt = avg_log2FC)

DoMultiBarHeatmap(imm_myelo_subset1, features = top_myelo_markers$gene, group.by = "cell_type_imm", additional.group.by = "tissue_type",additional.group.sort.by = "tissue_type", cols.use = list(tissue_type = use_colors), draw.lines = F) +
  scale_fill_viridis()
ggsave2("HeatMap_Myelo_Tumor.pdf", path = "output/fig3", width = 30, height = 40, units = "cm")
ggsave2("HeatMap_Myelo_Tumor.png", path = "output/fig3", width = 30, height = 40, units = "cm")



###only monocyte-derived myeloid cells

imm_myelo_subset2 <- subset(imm_myelo, cell_type_imm == "Myeloid")

Idents(imm_myelo_subset2) <- imm_myelo_subset2@meta.data$cell_type_imm_refined

myelo_markers <- FindAllMarkers(imm_myelo_subset2, only.pos = T, min.pct = 0.25, min.diff.pct = 0.25)

top_myelo_markers <- myelo_markers %>% group_by(cluster) %>% top_n(50, wt = avg_log2FC)

DoMultiBarHeatmap(imm_myelo_subset2, features = top_myelo_markers$gene, group.by = "cell_type_imm_refined", additional.group.by = "tissue_type",additional.group.sort.by = "tissue_type", cols.use = list(tissue_type = use_colors), draw.lines = F) +
  scale_fill_viridis()
ggsave2("HeatMap_Myelo_Tumor_Myeloid.pdf", path = "output/fig3", width = 30, height = 80, units = "cm")
ggsave2("HeatMap_Myelo_Tumor_Myeloid.png", path = "output/fig3", width = 30, height = 80, units = "cm")

```

```{r differential gene expression in lymphoid cells}

###all lymphoid cells

Idents(imm_lympho) <- imm_lympho@meta.data$cell_type_imm

markers_lympho <- FindAllMarkers(imm_lympho, only.pos = T, min.pct = 0.25, min.diff.pct = 0.25)

top_markers_lympho <- markers_lympho %>% group_by(cluster) %>% top_n(10, wt = avg_log2FC)

DoMultiBarHeatmap(imm_lympho, features = top_markers_lympho$gene, group.by = "cell_type_imm", additional.group.by = "tissue_type",additional.group.sort.by = "tissue_type", cols.use = list(tissue_type = use_colors), draw.lines = F) +
  scale_fill_viridis()
ggsave2("HeatMap_lympho_all.pdf", path = "output/fig3", width = 30, height = 35, units = "cm")
ggsave2("HeatMap_lympho_all.png", path = "output/fig3", width = 30, height = 35, units = "cm")

```

```{r T cell signatures}

###signatures according to Guo et al. 2018 https://www.nature.com/articles/s41591-018-0045-3

T_exhausted <- read_excel("data/CD8_T_cells_exhausted.xlsx", skip = 1)
cytotoxicity <- c("PRF1", "IFNG", "GNLY", "NKG7", "GZMB", "GZMA", "GZMH", "KLRK1", "KLRB1", "KLRD1", "CTSW", "CST7")
naiveness <- c("CCR7", "TCF7", "LEF1", "SELL")

T_cell_markers <- list(T_exhausted$GeneSymbol, cytotoxicity, naiveness)

imm_lympho <- AddModuleScore(imm_lympho, features = T_cell_markers, name = c("exhaustion", "cytotoxicity", "naiveness"), nbin = 12)

VlnPlot(imm_lympho, features = c("naiveness3", "cytotoxicity2", "exhaustion1"), pt.size = 0, group.by = "cell_type_imm_refined", ncol = 1, cols = use_colors)
#ggsave2("VlnPlot_cytotoxicity_exhaustion_imm.pdf", path = "output/fig3", width = 10, height = 20, units = "cm")

VlnPlot(imm_lympho, features = c("naiveness3", "cytotoxicity2", "exhaustion1"), pt.size = 0, group.by = "cell_type_imm", split.by = "tissue_type", ncol = 1)
ggsave2("VlnPlot_cytotoxicity_exhaustion_imm.pdf", path = "output/fig3", width = 10, height = 20, units = "cm")

```

```{r DC and monocyte subsets}

DotPlot(imm_myelo, features = c("IRF4", "CD1C", "IRF8", "THBD"), idents = "cDC", group.by = "cell_type_imm_refined") + 
  coord_flip() +
  scale_color_viridis()
ggsave2("DotPlot_cDC_subsets.pdf", path = "output/fig3", width = 10, height = 8, units = "cm")

DotPlot(imm_myelo, features = rev(c("CD14", "CD36", "S100A12", "ALOX5AP", "GFRA2", "NKG7", "PLVAP", "FCGR3A", "CXCR4", "HES4","C1QA")), idents = c("Monocyte"), group.by = "cell_type_imm_refined") + 
  coord_flip() +
  scale_color_viridis()
ggsave2("DotPlot_monocyte_subsets.pdf", path = "output/fig3", width = 13, height = 8, units = "cm")

DotPlot(imm_myelo, features = rev(c("APOE", "S100A8", "S100A9", "FABP4", "LGMN")), idents = "Myeloid", group.by = "cell_type_imm_refined") + 
  coord_flip() +
  scale_color_viridis()

```

```{r hallmark signatures}

VlnPlot(imm_myelo, features = c("HALLMARK_INFLAMMATORY_RESPONSE31",
                               "HALLMARK_INTERFERON_GAMMA_RESPONSE19",
                               "HALLMARK_TNFA_SIGNALING_VIA_NFKB1",
                               "HALLMARK_IL6_JAK_STAT3_SIGNALING7"), 
        group.by = "cell_type_imm", pt.size = 0, ncol = 1, split.by = "tissue_type")
ggsave2("VlnPlot_inflammatory.pdf", path = "output/fig3", width = 20, height = 30, units = "cm")

```

#cell type ingestion for comparison with reference dataset (lung adenocarcinoma)

```{r}

imm_anno$cell_type_imm <- as.character(imm_anno$cell_type_imm)

imm_anno_carcinoid <- subset(imm_anno, tissue_type == "Tumor")

SaveH5Seurat(imm_anno_carcinoid, filename = "loom_files/imm_anno_carcinoid.h5Seurat")

Convert("loom_files/imm_anno_carcinoid.h5Seurat", dest = "loom_files/imm_anno_carcinoid.h5ad")

```

#switch to python script for scanpy ingest and then come back to this R script

```{r}

cell_type_imm_new <- read.csv("ingest/mapped_carcinoid_imm/obs.csv")
rownames(cell_type_imm_new) <- cell_type_imm_new$X
cell_type_imm_new <- cell_type_imm_new %>% select(cell_type_imm)
cell_type_imm_new$cell_type_imm_new <- cell_type_imm_new$cell_type_imm
cell_type_imm_new$cell_type_imm <- NULL
imm_anno_carcinoid <- AddMetaData(imm_anno_carcinoid, metadata = cell_type_imm_new)

umap_imm_new <- read.csv("ingest/mapped_carcinoid_imm/obsm.csv")
rownames(umap_imm_new) <- rownames(cell_type_imm_new)
umap_imm_new$UMAP_1 <- umap_imm_new$X_umap1
umap_imm_new$UMAP_2 <- umap_imm_new$X_umap2
umap_imm_new <- umap_imm_new %>% select(UMAP_1, UMAP_2) %>% as.matrix()
imm_anno_carcinoid@reductions$umap@cell.embeddings <- umap_imm_new

```

```{r ingested cell type annotation}

use_colors2 <- c(
  Alveolar_Macrophages1 = "#6bAEd6",
  Alveolar_Macrophages2 = "#3182BD",
  Alveolar_Macrophages3 = "#08519C",
  CD14_Macrophages1= "#fff500",
  CD14_Macrophages2= "#FE9929",
  CD14_Macrophages3= "#EC7014",
  CD14_Macrophages4= "#CC4C02",
  CD14_Macrophages5= "#8C2D04",
  Macrophages_Proliferating= "#E31A1C",
  Monocytes= "#FA9FB5",
  Myeloid_Dendritic= "#DD3497",
  Plasmacytoid_Dendritic= "#7A0177",
  T_conv1= "#c2e699",
  T_conv2= "#78c679",
  T_reg= "#006837",
  T_CD8_1= "#bcbddc",
  T_CD8_2= "#9e9ac8",
  T_CD8_3= "#807dba",
  T_CD8_Proliferating= "#6a51a3",
  NK_cells= "#4a1486",
  B_cells= "#969696",
  Plasma= "#636363",
  Mast= "#252525")

imm_anno_carcinoid@meta.data$cell_type_imm_new <- ordered(imm_anno_carcinoid@meta.data$cell_type_imm_new, levels = c("Alveolar_Macrophages1",
                                                                                         "Alveolar_Macrophages2",
                                                                                         "Alveolar_Macrophages3",
                                                                                         "CD14_Macrophages1",
                                                                                         "CD14_Macrophages2",
                                                                                         "CD14_Macrophages3",
                                                                                         "CD14_Macrophages4",
                                                                                         "CD14_Macrophages5",
                                                                                         "Macrophages_Proliferating",
                                                                                         "Monocytes",
                                                                                         "Myeloid_Dendritic",
                                                                                         "Plasmacytoid_Dendritic",
                                                                                         "Mast",
                                                                                         "T_conv1",
                                                                                         "T_conv2",
                                                                                         "T_reg",
                                                                                         "T_CD8_1",
                                                                                         "T_CD8_2",
                                                                                         "T_CD8_3",
                                                                                         "T_CD8_Proliferating",
                                                                                         "NK_cells",
                                                                                         "B_cells",
                                                                                         "Plasma"))

imm_lympho <- subset(imm_anno_carcinoid, subset = cell_type_imm_new %in% c("T_conv1",
                                                           "T_conv2",
                                                           "T_reg",
                                                           "T_CD8_1",
                                                           "T_CD8_2",
                                                           "T_CD8_3",
                                                           "T_CD8_Proliferating",
                                                           "NK_cells",
                                                           "B_cells",
                                                           "Plasma"))

imm_lympho <- ScaleData(imm_lympho)

imm_myelo <- subset(imm_anno_carcinoid, subset = cell_type_imm_new %in% c("Alveolar_Macrophages1",
                                                          "Alveolar_Macrophages2",
                                                          "Alveolar_Macrophages3",
                                                          "CD14_Macrophages1",
                                                          "CD14_Macrophages2",
                                                          "CD14_Macrophages3",
                                                          "CD14_Macrophages4",
                                                          "CD14_Macrophages5",
                                                          "Macrophages_Proliferating",
                                                          "Monocytes",
                                                          "Myeloid_Dendritic",
                                                          "Plasmacytoid_Dendritic",
                                                          "Mast"))

imm_myelo <- ScaleData(imm_myelo)

lympho_counts <- FetchData(imm_lympho, vars = c("tissue_type", "cell_type_imm_new", "sample_id", "patient_id")) %>%  
  mutate(tissue_type = factor(tissue_type, levels = c("Tumor", "Normal")))

lympho_counts_tbl <- lympho_counts %>%
  dplyr::count(cell_type_imm_new, patient_id, tissue_type)
#write_csv(lympho_counts_tbl, path = "output/fig5/lympho_counts.csv")

myelo_counts <- FetchData(imm_myelo, vars = c("tissue_type", "cell_type_imm_new", "sample_id", "patient_id")) %>%  
  mutate(tissue_type = factor(tissue_type, levels = c("Tumor", "Normal"))) 

myelo_counts_tbl <- myelo_counts %>%
  dplyr::count(cell_type_imm_new, patient_id, tissue_type)
#write_csv(myelo_counts_tbl, path = "output/fig5/myelo_counts.csv")


ggplot(data = lympho_counts, aes(x = tissue_type, fill = cell_type_imm_new)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors2) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_lympho.pdf", path = "output/ingested_annotation", width = 20, height = 5, units = "cm")

ggplot(data = myelo_counts, aes(x = tissue_type, fill = cell_type_imm_new)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors2) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_myelo.pdf", path = "output/ingested_annotation", width = 20, height = 5, units = "cm")

lympho_counts %>%
  filter(tissue_type == "Tumor") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm_new)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors2) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_lympho_patients.pdf", path = "output/ingested_annotation", width = 20, height = 5, units = "cm")

myelo_counts %>%
  filter(tissue_type == "Tumor") %>%
  ggplot(aes(x = sample_id, fill = cell_type_imm_new)) +
  geom_bar(position = "fill") +
  scale_fill_manual(values = use_colors2) +
  coord_flip() +
  scale_y_reverse()
ggsave2("BarPlot_cell_type_imm_myelo_patients.pdf", path = "output/ingested_annotation", width = 20, height = 5, units = "cm")

DimPlot(imm_anno_carcinoid, group.by = "cell_type_imm", cols = use_colors)
ggsave2("UMAP_cell_type_imm.pdf", path = "output/ingested_annotation", width = 25, height = 15, units = "cm")
ggsave2("UMAP_cell_type_imm.png", path = "output/ingested_annotation", width = 25, height = 15, units = "cm")

DimPlot(imm_anno_carcinoid, group.by = "cell_type_imm_new", cols = use_colors2)
ggsave2("UMAP_cell_type_imm_new.pdf", path = "output/ingested_annotation", width = 25, height = 15, units = "cm")
ggsave2("UMAP_cell_type_imm_new.png", path = "output/ingested_annotation", width = 25, height = 15, units = "cm")

```

```{r}

cell_types <- FetchData(imm_anno_carcinoid, c("cell_type_imm", "cell_type_imm_new"))

cell_types <- cell_types %>% make_long(cell_type_imm, cell_type_imm_new)

ggplot(cell_types, aes(x = x, 
               next_x = next_x, 
               node = node, 
               next_node = next_node,
               fill = factor(node),
               label = node)) +
  geom_sankey() +
  geom_sankey_label() +
  scale_fill_manual(values = c(use_colors, use_colors2)) +
  theme_sankey() +
  theme(legend.position = "none")
ggsave2("sankey_imm.pdf", path = "output/ingested_annotation", width = 15, height = 15, units = "cm")

```
