==> 07_Carcinoid_CellPhoneDB.Rmd <==

```{r}

library(Seurat)
library(readr)
library(tidyverse)

```

```{r}

epi_anno <- readRDS("seurat_objects/epi_anno.RDS")
epi_tumor <- subset(epi_anno, subset = cluster_type == "Tumor" & tissue_type == "Tumor")

imm_anno <- readRDS("seurat_objects/imm_anno.RDS")
imm_subset <- subset(imm_anno, subset = cell_type_imm %in% c("Myeloid",
                                                             "cDC",
                                                             "Monocyte")
                     & tissue_type == "Tumor")

str_anno <- readRDS("seurat_objects/str_anno.RDS")
str_subset <- subset(str_anno, subset = cell_type_str %in% c("Myofibro",
                                                             "Muscle_Vascular",
                                                             "Pericyte",
                                                             "Endothelial_Bronchial")
                     & tissue_type == "Tumor")

```

```{r}

epi_tumor$cell_type <- epi_tumor$cell_type_epi
imm_subset$cell_type <- imm_subset$cell_type_imm
str_subset$cell_type <- str_subset$cell_type_str

subset_combined <- merge(epi_tumor, c(imm_subset, str_subset))

remove(epi_anno)
remove(epi_tumor)
remove(imm_anno)
remove(imm_subset)
remove(str_anno)
remove(str_subset)

```

```{r}

# generating meta file with cell type info

meta_data <- FetchData(subset_combined, vars = "cell_type")

meta_data$cell_id <- rownames(meta_data)

write_csv(meta_data, file = "output/CellPhoneDB/meta_data.csv")

```

```{r}

count_raw <- GetAssayData(object = subset_combined, assay = "RNA", slot = "counts")

remove(subset_combined)

count_raw1 <- count_raw[,1:5000]
count_raw2 <- count_raw[,5001:ncol(count_raw)]

count_norm1 <- apply(count_raw1, 2, function(x) (x/sum(x))*10000)
count_norm2 <- apply(count_raw2, 2, function(x) (x/sum(x))*10000)

count_norm <- cbind(count_norm1, count_norm2)

count_norm1 <- NULL
count_norm2 <- NULL
count_raw1 <- NULL
count_raw2 <- NULL

write.csv(count_norm, "output/CellPhoneDB/count_norm.csv")

write.csv(count_raw, "output/CellPhoneDB/count_raw.csv")

```
